Remove old files and rename new files
