+++ sessionize/lib/src/commonMain/kotlin/co/touchlab/sessionize/db/SponsorGroupDbItem.kt
@@ -1,0 +1,5 @@
+package co.touchlab.sessionize.db
+
+import co.touchlab.droidcon.db.Sponsor
+
+data class SponsorGroupDbItem(val groupName: String, val sponsors: List<Sponsor>)
